// src/types/character.ts
export enum Stats {
  Strength = 'Strength',
  Dexterity = 'Dexterity',
  Constitution = 'Constitution',
  Intelligence = 'Intelligence',
  Wisdom = 'Wisdom',
  Charisma = 'Charisma',
}

export enum Lvl {
  One = 'One',
  Two = 'Two',
  Three = 'Three',
  Four = 'Four',
  Five = 'Five',
  Six = 'Six',
  Seven = 'Seven',
  Eight = 'Eight',
  Nine = 'Nine',
  Ten = 'Ten',
  Eleven = 'Eleven',
  Twelve = 'Twelve',
  Thirteen = 'Thirteen',
  Fourteen = 'Fourteen',
  Fifteen = 'Fifteen',
  Sixteen = 'Sixteen',
  Seventeen = 'Seventeen',
  Eighteen = 'Eighteen',
  Nineteen = 'Nineteen',
  Twenty = 'Twenty',
}

export const Experience = {
  [Lvl.One]: 0,
  [Lvl.Two]: 300,
  [Lvl.Three]: 900,
  [Lvl.Four]: 2700,
  [Lvl.Five]: 6500,
  [Lvl.Six]: 14000,
  [Lvl.Seven]: 23000,
  [Lvl.Eight]: 34000,
  [Lvl.Nine]: 48000,
  [Lvl.Ten]: 64000,
  [Lvl.Eleven]: 85000,
  [Lvl.Twelve]: 100000,
  [Lvl.Thirteen]: 120000,
  [Lvl.Fourteen]: 140000,
  [Lvl.Fifteen]: 165000,
  [Lvl.Sixteen]: 195000,
  [Lvl.Seventeen]: 225000,
  [Lvl.Eighteen]: 265000,
  [Lvl.Nineteen]: 305000,
  [Lvl.Twenty]: 355000,
};

export const Proficiency = {
  [Lvl.One]: 2,
  [Lvl.Two]: 2,
  [Lvl.Three]: 2,
  [Lvl.Four]: 2,
  [Lvl.Five]: 3,
  [Lvl.Six]: 3,
  [Lvl.Seven]: 3,
  [Lvl.Eight]: 3,
  [Lvl.Nine]: 4,
  [Lvl.Ten]: 4,
  [Lvl.Eleven]: 4,
  [Lvl.Twelve]: 4,
  [Lvl.Thirteen]: 5,
  [Lvl.Fourteen]: 5,
  [Lvl.Fifteen]: 5,
  [Lvl.Sixteen]: 5,
  [Lvl.Seventeen]: 6,
  [Lvl.Eighteen]: 6,
  [Lvl.Nineteen]: 6,
  [Lvl.Twenty]: 6,
};

export enum Class {
  Artificer = 'Artificer',
  Barbarian = 'Barbarian',
  Bard = 'Bard',
  Cleric = 'Cleric',
  Druid = 'Druid',
  Fighter = 'Fighter',
  Monk = 'Monk',
  Paladin = 'Paladin',
  Ranger = 'Ranger',
  Rogue = 'Rogue',
  Sorcerer = 'Sorcerer',
  Warlock = 'Warlock',
  Wizard = 'Wizard',
}

export default { Stats, Lvl };